Split feed module off the API module
